+++ src/web/constants/brokenImageGlyph.ts
@@ -1,8 +1,7 @@
-// SVG path of the broken-image glyph, shared between the editor's image
-// node view (rendered as an inline <svg>).
+// SVG path of the broken-image glyph, shared between the editor's
+// image node view (rendered as an inline <svg>).
 export const BROKEN_IMAGE_PATH_D =
   'M200,840Q167,840 143.5,816.5Q120,793 120,760L120,200Q120,167 143.5,143.5Q167,120 200,120L760,120Q793,120 816.5,143.5Q840,167 840,200L840,760Q840,793 816.5,816.5Q793,840 760,840L200,840ZM240,503L400,343L560,503L720,343L760,383L760,200Q760,200 760,200Q760,200 760,200L200,200Q200,200 200,200Q200,200 200,200L200,463L240,503ZM200,760L760,760Q760,760 760,760Q760,760 760,760L760,496L720,456L560,616L400,456L240,616L200,576L200,760Q200,760 200,760Q200,760 200,760ZM200,760L200,760Q200,760 200,760Q200,760 200,760L200,496L200,576L200,463L200,383L200,200Q200,200 200,200Q200,200 200,200L200,200Q200,200 200,200Q200,200 200,200L200,463L200,463L200,576L200,576L200,760Q200,760 200,760Q200,760 200,760Z';
 
-// Glyph as a `mask`-ready data URI, consumed via the
-// `--et-broken-image-glyph` custom property in EnrichedText.css.
+// Glyph as a `mask`-ready data URI, used in EnrichedText.css.
 export const BROKEN_IMAGE_GLYPH_MASK = `url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 960 960'%3E%3Cpath d='${BROKEN_IMAGE_PATH_D}'/%3E%3C/svg%3E")`;
